requirements.txt
# MTS MultAgent Employee Monitoring System - Dependencies

# Core async framework
asyncio-mqtt>=0.13.0
aiofiles>=23.2.1
aiohttp>=3.8.6

# HTTP Client & API
httpx>=0.24.1
requests>=2.31.0
urllib3>=2.0.7

# Data processing & analysis
pandas>=2.1.4
numpy>=1.24.4
python-dateutil>=2.8.2

# Document processing
python-docx>=0.8.11
PyPDF2>=3.0.1
openpyxl>=3.1.2
xlrd>=2.0.1

# Configuration management
PyYAML>=6.0.1
python-dotenv>=1.0.0

# Logging & monitoring
structlog>=23.2.0
psutil>=5.9.6

# LLM Integration
openai>=1.3.7
tiktoken>=0.5.2

# Scheduling
schedule>=1.2.0
apscheduler>=3.10.4

# Testing framework
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-mock>=3.12.0
pytest-cov>=4.1.0

# Development tools
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.1
isort>=5.12.0

# API Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Security & validation
pydantic>=2.5.0
cryptography>=41.0.7
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# Database & storage (optional future enhancements)
sqlalchemy>=2.0.23
alembic>=1.13.0
redis>=5.0.1

# Monitoring & observability
prometheus-client>=0.19.0
grafana-api>=1.0.3

# Utilities
click>=8.1.7
rich>=13.7.0
typer>=0.9.0
pathlib2>=2.3.7; python_version < "3.4"

# Development server
watchdog>=3.0.0

# Performance profiling
py-spy>=0.3.14
memory-profiler>=0.61.0

# Environment-specific dependencies (uncomment as needed)
# Production database drivers
# psycopg2-binary>=2.9.9
# pymysql>=1.1.0

# Cloud provider SDKs (for future integrations)
# boto3>=1.29.7
# google-cloud-storage>=2.10.0
# azure-storage-blob>=12.19.0

# Advanced analytics (optional)
# scikit-learn>=1.3.2
# matplotlib>=3.8.2
# seaborn>=0.13.0
# plotly>=5.17.0

# Notification systems (optional)
# slack-sdk>=3.23.0
# python-telegram-bot>=20.6
# smtplib2>=0.2.1
